requirements.txt
Flask 
numpy 
pandas 
autopep8 
Flask-RESTful 
matplotlib 
pathlib 
requests 
joblib 
opencv-python 
cmake 
sklearn 
face-recognition 
dlib 
pillow
uwsgi
flask_cors
tensorflow